--- examples/data_efficient_nopt/models/utils.py
@@ -1,51 +1,0 @@
-import paddle.nn.functional as F
-
-
-def add_padding(x, num_pad):
-    if max(num_pad) > 0:
-        res = F.pad(x, (num_pad[0], num_pad[1]), "constant", 0)
-    else:
-        res = x
-    return res
-
-
-def add_padding2(x, num_pad1, num_pad2):
-    if max(num_pad1) > 0 or max(num_pad2) > 0:
-        res = F.pad(
-            x, (num_pad2[0], num_pad2[1], num_pad1[0], num_pad1[1]), "constant", 0.0
-        )
-    else:
-        res = x
-    return res
-
-
-def remove_padding(x, num_pad):
-    if max(num_pad) > 0:
-        res = x[..., num_pad[0] : -num_pad[1]]
-    else:
-        res = x
-    return res
-
-
-def remove_padding2(x, num_pad1, num_pad2):
-    if max(num_pad1) > 0 or max(num_pad2) > 0:
-        res = x[..., num_pad1[0] : -num_pad1[1], num_pad2[0] : -num_pad2[1]]
-    else:
-        res = x
-    return res
-
-
-def _get_act(act):
-    if act == "tanh":
-        func = F.tanh
-    elif act == "gelu":
-        func = F.gelu
-    elif act == "relu":
-        func = F.relu_
-    elif act == "elu":
-        func = F.elu_
-    elif act == "leaky_relu":
-        func = F.leaky_relu_
-    else:
-        raise ValueError(f"{act} is not supported")
-    return func
